% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lafit.R
\name{lafit}
\alias{lafit}
\title{The LA and LI Distribution}
\usage{
lafit(x)
}
\arguments{
\item{x}{data}

\item{alpha}{parameter 1.}

\item{gama}{parameter 2.}

\item{L}{parameter 3.}
}
\description{
Fit function for the LA distribution.
}
\examples{

lafit(rla(n=1000,alpha=-10,gama=1.5,L=4))

}
