% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifit.R
\name{lifit}
\alias{lifit}
\title{The LA and LI Distribution}
\usage{
lifit(x)
}
\arguments{
\item{x}{data}

\item{alpha}{parameter 1.}

\item{gama}{parameter 2.}

\item{L}{parameter 3.}
}
\description{
Fit function for the LI distribution.
}
\examples{

lifit(rli(n=1000,alpha=-10,gama=1.5,L=4))

}
