% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pli.R
\name{pli}
\alias{pli}
\title{The LA and LI distribution}
\usage{
pli(q, alpha, gama, L, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{alpha}{parameter 1.}

\item{gama}{parameter 2.}

\item{L}{parameter 3.}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\description{
Cumulative density function of the LI distribution.
}
\examples{

pli(q=0.3511,alpha=-10,gama=1.5,L=4)

}
