% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qla.R
\name{qla}
\alias{qla}
\title{The LA and LI distribution}
\usage{
qla(p, alpha, gama, L)
}
\arguments{
\item{p}{vector of probabilities.}

\item{alpha}{parameter 1.}

\item{gama}{parameter 2.}

\item{L}{parameter 3.}
}
\description{
Quantile function of the LA distribution.
}
\examples{

qla(p=0.4,alpha=-2,gama=2,L=3)

}
