% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qli.R
\name{qli}
\alias{qli}
\title{The LA and LI distribution}
\usage{
qli(p, alpha, gama, L)
}
\arguments{
\item{p}{vector of probabilities.}

\item{alpha}{parameter 1.}

\item{gama}{parameter 2.}

\item{L}{parameter 3.}
}
\description{
Quantile function of the LI distribution.
}
\examples{

qli(p=0.4,alpha=-2,gama=2,L=3)

}
