% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rla.R
\name{rla}
\alias{rla}
\title{The LA and LI distribution}
\usage{
rla(n, alpha, gama, L)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{alpha}{parameter 1.}

\item{gama}{parameter 2.}

\item{L}{parameter 3.}
}
\description{
Random generation for the LA distribution.
}
\examples{

rla(n=1000,alpha=-2,gama=2,L=3)

}
