% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rli.R
\name{rli}
\alias{rli}
\title{The LA and LI distribution}
\usage{
rli(n, alpha, gama, L)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{alpha}{parameter 1.}

\item{gama}{parameter 2.}

\item{L}{parameter 3.}
}
\description{
Random generation for the LI distribution.
}
\examples{

rli(n=1000,alpha=-10,gama=1.5,L=1.5)

}
